fix int in db

+++ src/achievement/models.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 from fastapi_users_db_sqlalchemy import SQLAlchemyBaseUserTable
-from sqlalchemy import Table, Column, Integer, String, TIMESTAMP, ForeignKey, JSON, Boolean, Text
+from sqlalchemy import Table, Column, Integer, String, TIMESTAMP, ForeignKey, JSON, Boolean, Text, BIGINT
 
 from src.user_profile.models import user
 from src.database import Base, metadata
@@ -17,7 +17,7 @@ achievement = Table(
 user_x_achievement = Table(
     'user_x_achievement', metadata,
     Column('id', Integer, primary_key=True, autoincrement=True),
-    Column('user_id', Integer, ForeignKey('user.id'), nullable=False),
+    Column('user_id', BIGINT, ForeignKey('user.id'), nullable=False),
     Column('achievement_id', Integer, ForeignKey('achievement.id'), nullable=False),
     Column('date', TIMESTAMP, nullable=False)
 )
