+++ src/market/models.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 from fastapi_users_db_sqlalchemy import SQLAlchemyBaseUserTable
-from sqlalchemy import Table, Column, Integer, String, TIMESTAMP, ForeignKey, JSON, Boolean, Text
+from sqlalchemy import Table, Column, Integer, String, TIMESTAMP, ForeignKey, JSON, Boolean, Text, BIGINT
 
 from src.user_profile.models import user
 from src.database import Base, metadata
@@ -21,7 +21,7 @@ product = Table(
 user_x_product = Table(
     'user_x_product', metadata,
     Column('id', Integer, primary_key=True),
-    Column('user_id', Integer, ForeignKey('user.id'), nullable=False),
+    Column('user_id', BIGINT, ForeignKey('user.id'), nullable=False),
     Column('product_id', Integer, ForeignKey('product.id'), nullable=False),
     Column('club_id', Integer, ForeignKey('club.id'), nullable=False),
     Column('date', TIMESTAMP, nullable=False),
